Changed from create-react-app to vite
This involved recreating the whole thing, deleting all the files
from this one, and then copying the vite version's files into this one.
So there are a lot of changes. There's some messiness left over with
the merging of the CSSes, and some of the template code from the vite
starter app is still here. But all the functionality is complete.